Engine Folder Openning + Compiler warning fixes (#155)
* Engine Folder Openning + Compiler warning fixes

* Drag Drop multiple files.

* able to drop on tree.

* removed redundant null check.

* copy, cut, paste with Ctrl+C, Ctrl+V

* bug fix.

* Left click copy paste works.

* no need g_current entry.

* esc discard selection, shift instead of ctrl.

* bug, feedback fixes.

* more comments.

* last feedback fixes. ctrl deselect, seperate FolderView.cpp

* cutting, copying highlight colors. and fix right click behaviour

---------

+++ Editor/PluginWindow.cpp
@@ -33,7 +33,7 @@ namespace ToolKit
       m_screenResolutions.push_back(Vec2(500.0f, 500.0f));
     }
 
-    void PluginWindow::RemoveResolutionName(int index)
+    void PluginWindow::RemoveResolutionName(size_t index)
     {
       bool canRemove = index > 0 || index < m_screenResolutions.size();
